Fix output when a package path is provided
Providing a full path to the interface is useful in cases where the
interface can't be guessed just from the package name and interface
name, however this is not currently accounted for when the output is
printed.

This modifies the logic to use the id we get from `findInterface`
instead.